Refactor WebSocket connection management: Introduce WebSocketConnector class
- Replaced ConnectionManager with WebSocketConnector for improved connection handling.
- Enhanced connection state management with new ConnectionState enum.
- Implemented retry logic for connection attempts with configurable options.
- Updated message sending methods to handle both string and binary messages.
- Introduced inflight request tracking and queuing for better request management.
- Added support for multi-host connection URLs.
- Removed legacy WebSocketConnectionPool and associated methods for cleaner architecture.
- Updated unit tests to reflect changes in connection management and ensure functionality.
